feat(GHB-01): create first version of the project
Github Blog is a project to show the latest posts from a Github created
inside of Issues page with specific tags, in this case is "article".

The ideia is finish the project in 30 days and create all following
using Pomodoro timer to increase productivity and focus. And show how
much time you need to spent to build all this.

We have issues to apply, but the challenge is finish the things
necessary to deliver this project.

Resolves: GHB-01, GHB-04, GHB-5, GHB-06, GHB-07, GHB-08, GHB-09, GHB-10,
GHB-11, GHB-12, GHB-13, GHB-14, GHB-15